+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.213] - 2026-05-22
+
+### Fixed
+- Prevent Windsurf autopilot from using the sidebar toggle command that can close the chat panel.
+- Bundle koru autopilot VSIX 0.1.30.
+
 ## [0.1.211] - 2026-05-22
 
 ### Fixed